requirements.txt
pytest
requests
allure-pytest
requests-toolbelt